+++ R/zzz.R
@@ -1,8 +1,9 @@
 #' @importFrom cli cli_alert_warning cli_progress_step cli_progress_done
 #' @importFrom cli col_yellow
+#' @importFrom graphics arrows box points text
 #' @importFrom dplyr across all_of
 #' @importFrom dplyr filter join_by left_join mutate rename select
-#' @importFrom stats formula
+#' @importFrom stats as.formula kmeans formula
 #' @importFrom vegan cascadeKM ordiR2step rda RsquareAdj scores
 #'
 NULL