+++ NEWS.md
@@ -1,3 +1,10 @@
+# classEcoFor (0.2.0)
+
+* Complete reproduction of Chalumeau et al. 
+* Add POE data.
+* Add `triplot_rda()`
+* Use `FLAT_SLOPE` instead of all `SITUATION` values.
+
 # classEcoFor (0.1.0)
 
 * Minimal code to partially reproduce Chalumeau et al.